+++ templates/Models/Location.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Support\Facades\DB;
+use Hamatoma\Laraknife\ViewHelpers;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+
+class Location extends Model
+{
+    use HasFactory;
+    protected $table = 'locations';
+    protected $fillable = [
+        'country',
+        'zip',
+        'city',
+        'street',
+        'additional',
+        'info',
+        'priority',
+        'owner_id'
+    ];
+}