+++ templates/database/migrations/2024_11_30_211027_create_locations_table.php
@@ -1,0 +1,35 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::create('locations', function (Blueprint $table) {
+            $table->id();
+            $table->timestamps();
+            $table->string('country', 5);
+            $table->string('zip', 15);
+            $table->string('city', 128);
+            $table->string('street', 128);
+            $table->string('additional')->nullable();
+            $table->string('info')->nullable();
+            $table->integer('priority');
+            $table->foreignId('owner_id')->nullable()->references('id')->on('users');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::dropIfExists('locations');
+    }
+};